chore: move target_node runtime into runtime directory
