+++ crates/pack-tests/tests/snapshot/runtime/target_node_runtime/input/index.js
@@ -1,0 +1,1 @@
+console.log('Hello, world!')
