+++ crates/pack-tests/tests/snapshot/runtime/target_node_runtime/output/_project__runtime_target_node_runtime_input_index_6a9aad93.js
@@ -1,0 +1,8 @@
+module.exports = [
+"[project]/runtime/target_node_runtime/input/index.js [server] (ecmascript)", ((__turbopack_context__, module, exports) => {
+
+console.log('Hello, world!');
+}),
+];
+
+//# sourceMappingURL=_project__runtime_target_node_runtime_input_index_6a9aad93.js.map
