+++ crates/pack-tests/tests/snapshot/runtime/target_node_runtime/output/main.js
@@ -1,0 +1,4 @@
+var R=require("./_turbopack__runtime.js")("main.js")
+R.c("_project__runtime_target_node_runtime_input_index_6a9aad93.js")
+R.m("[project]/runtime/target_node_runtime/input/index.js [server] (ecmascript)")
+module.exports=R.m("[project]/runtime/target_node_runtime/input/index.js [server] (ecmascript)").exports
